requirements.txt
huggingface_hub
numpy
Pillow
matplotlib
scipy
scikit-image
tifffile
scikit-learn
opencv-python
torch
torchvision
h5py
tqdm